refactor: split the assistant chat operations into a module
